Refs #20815: Apply Mario's suggestions
